Some changes to reproduce better my intervention...
Hi Clinton. Reviewing the notes for the annual report today I arrive to this one, so I made some changes in the last part of the notes that I believe reflect better the intervention I made. 

As far as is a change in the notes, may be I need to send the corrections to the list?
If not, just merge, please.